+++ np.Templating/lib/support/modules/helpersModule.js
@@ -26,7 +26,7 @@ import { getFormattedTime, getISOWeekAndYear, getISOWeekString, hyphenatedDate, 
 
 import { getNPWeekData } from '@helpers/NPdateTime'
 
-import { parseJSON5, semverVersionToNumber, getRandomUUID } from '@helpers/general'
+import { parseJSON5, semverVersionToNumber, getRandomUUID, createOpenOrDeleteNoteCallbackUrl } from '@helpers/general'
 
 import { getNote, removeSection } from '@helpers/note'
 
@@ -101,6 +101,7 @@ const helpers = {
   parseJSON5,
   semverVersionToNumber,
   getRandomUUID,
+  createOpenOrDeleteNoteCallbackUrl,
 
   // Note-related helpers
   getNote,